ATO-1727: Move GlobalLogoutMessage to orchestration-shared
- We would like to send audit events from inside GlobalLogoutService. This is because if no sessions were found for the icsid, then we do not want to send an audit event.
