+++ integration-tests/src/test/java/uk/gov/di/authentication/api/GlobalLogoutIntegrationTest.java
@@ -16,8 +16,8 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import uk.gov.di.authentication.oidc.entity.GlobalLogoutMessage;
 import uk.gov.di.authentication.oidc.lambda.GlobalLogoutHandler;
+import uk.gov.di.orchestration.shared.entity.GlobalLogoutMessage;
 import uk.gov.di.orchestration.shared.entity.OrchClientSessionItem;
 import uk.gov.di.orchestration.shared.entity.OrchSessionItem;
 import uk.gov.di.orchestration.sharedtest.basetest.IntegrationTest;
