+++ oidc-api/src/test/java/uk/gov/di/authentication/oidc/validators/GlobalLogoutValidatorTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import uk.gov.di.authentication.oidc.entity.GlobalLogoutMessage;
 import uk.gov.di.authentication.oidc.exceptions.GlobalLogoutValidationException;
+import uk.gov.di.orchestration.shared.entity.GlobalLogoutMessage;
 
 import java.util.stream.Stream;
